dag validation test added
