+++ tests/dag_validation_test/google_spreadsheet_pipeline_dag_validation_test.py
@@ -4,17 +4,9 @@ from dags.google_spreadsheet_pipeline_controller import (
     DAG_ID as CONTROLLER_DAG_ID,
     TARGET_DAG_ID
 )
-
-
-def dag_should_contain_named_tasks(dagbag, dag_id, task_list):
-    dag = dagbag.get_dag(dag_id)
-    tasks = dag.tasks
-    task_ids = list(map(lambda task: task.task_id, tasks))
-    task_ids.sort()
-    expected_ids = task_list
-    expected_ids.sort()
-
-    assert task_ids == expected_ids
+from tests.dag_validation_test import (
+    dag_should_contain_named_tasks
+)
 
 
 def test_dags_should_contain_one_task(dagbag):
